+++ docs/source/changelog.rst
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 Added
 ^^^^^
 
+* ``long_running`` example.
 * Support for pluggable message |Encoders|.